% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepback.R
\name{stepback}
\alias{stepback}
\title{stepback.Rd}
\usage{
stepback(y = y, d = d, alpha = 0.05)
}
\arguments{
\item{y}{dependent variable for selection}

\item{d}{select the independent variables}
}
\description{
When choosing 'd', and d is a dataframe, then the entire dataframe can be used (look at example), although when this is the case you have to make sure the dependent variable is de-selected
}
\examples{
data(weight)
stepback(weight$wgt, weight[, -1], alpha = 0.2)
}
